Create separate test class to better test and highlight issues
