+++ src/test/java/org/openrewrite/java/logging/slf4j/JulToSlf4jTest.java
@@ -353,86 +353,6 @@ class JulToSlf4jTest implements RewriteTest {
     }
 
     @Test
-    void parametrizedLoggerCallsIsLoggable() {
-        rewriteRun(
-          // language=java
-          java(
-            """
-              import java.util.logging.Level;
-              import java.util.logging.Logger;
-
-              class Test {
-                  void method(Logger logger, String param1, String param2) {
-                      logger.log(Level.FINEST, "FINEST Log entry, param1: {0}", param1);
-                      logger.log(Level.FINEST, "FINEST Log entry, param1: {0}, param2: {1}, etc", new Object[]{param1, param2});
-                      logger.log(Level.FINEST, "FINEST Log entry, param1: {0}, param2: {1}, etc", new String[]{param1, param2});
-
-                      logger.log(Level.FINER, "FINER Log entry, param1: {0}", param1);
-                      logger.log(Level.FINER, "FINER Log entry, param1: {0}, param2: {1}, etc", new Object[]{param1, param2});
-                      logger.log(Level.FINER, "FINER Log entry, param1: {0}, param2: {1}, etc", new String[]{param1, param2});
-
-                      logger.log(Level.FINE, "FINE Log entry, param1: {0}", param1);
-                      logger.log(Level.FINE, "FINE Log entry, param1: {0}, param2: {1}, etc", new Object[]{param1, param2});
-                      logger.log(Level.FINE, "FINE Log entry, param1: {0}, param2: {1}, etc", new String[]{param1, param2});
-
-                      logger.log(Level.CONFIG, "CONFIG Log entry, param1: {0}", param1);
-                      logger.log(Level.CONFIG, "CONFIG Log entry, param1: {0}, param2: {1}, etc", new Object[]{param1, param2});
-                      logger.log(Level.CONFIG, "CONFIG Log entry, param1: {0}, param2: {1}, etc", new String[]{param1, param2});
-
-                      logger.log(Level.INFO, "INFO Log entry, param1: {0}", param1);
-                      logger.log(Level.INFO, "INFO Log entry, param1: {0}, param2: {1}, etc", new Object[]{param1, param2});
-                      logger.log(Level.INFO, "INFO Log entry, param1: {0}, param2: {1}, etc", new String[]{param1, param2});
-
-                      logger.log(Level.WARNING, "WARNING Log entry, param1: {0}", param1);
-                      logger.log(Level.WARNING, "WARNING Log entry, param1: {0}, param2: {1}, etc", new Object[]{param1, param2});
-                      logger.log(Level.WARNING, "WARNING Log entry, param1: {0}, param2: {1}, etc", new String[]{param1, param2});
-
-                      logger.log(Level.SEVERE, "SEVERE Log entry, param1: {0}", param1);
-                      logger.log(Level.SEVERE, "SEVERE Log entry, param1: {0}, param2: {1}, etc", new Object[]{param1, param2});
-                      logger.log(Level.SEVERE, "SEVERE Log entry, param1: {0}, param2: {1}, etc", new String[]{param1, param2});
-                  }
-              }
-              """,
-            """
-              import org.slf4j.Logger;
-
-              class Test {
-                  void method(Logger logger, String param1, String param2) {
-                      logger.trace("FINEST Log entry, param1: {}", param1);
-                      logger.trace("FINEST Log entry, param1: {}, param2: {}, etc", param1, param2);
-                      logger.trace("FINEST Log entry, param1: {}, param2: {}, etc", param1, param2);
-
-                      logger.trace("FINER Log entry, param1: {}", param1);
-                      logger.trace("FINER Log entry, param1: {}, param2: {}, etc", param1, param2);
-                      logger.trace("FINER Log entry, param1: {}, param2: {}, etc", param1, param2);
-
-                      logger.debug("FINE Log entry, param1: {}", param1);
-                      logger.debug("FINE Log entry, param1: {}, param2: {}, etc", param1, param2);
-                      logger.debug("FINE Log entry, param1: {}, param2: {}, etc", param1, param2);
-
-                      logger.info("CONFIG Log entry, param1: {}", param1);
-                      logger.info("CONFIG Log entry, param1: {}, param2: {}, etc", param1, param2);
-                      logger.info("CONFIG Log entry, param1: {}, param2: {}, etc", param1, param2);
-
-                      logger.info("INFO Log entry, param1: {}", param1);
-                      logger.info("INFO Log entry, param1: {}, param2: {}, etc", param1, param2);
-                      logger.info("INFO Log entry, param1: {}, param2: {}, etc", param1, param2);
-
-                      logger.warn("WARNING Log entry, param1: {}", param1);
-                      logger.warn("WARNING Log entry, param1: {}, param2: {}, etc", param1, param2);
-                      logger.warn("WARNING Log entry, param1: {}, param2: {}, etc", param1, param2);
-
-                      logger.error("SEVERE Log entry, param1: {}", param1);
-                      logger.error("SEVERE Log entry, param1: {}, param2: {}, etc", param1, param2);
-                      logger.error("SEVERE Log entry, param1: {}, param2: {}, etc", param1, param2);
-                  }
-              }
-              """
-          )
-        );
-    }
-
-    @Test
     void staticFinalLoggerIsStaticFinal() {
         //language=java
         rewriteRun(
